feat: can decide how many page to fetch
