+++ example/demo.js
@@ -1,14 +1,15 @@
 const crawler = require('../index');
 
-async function runCrawler() {
+(async () => {
   // *** Initialize ***
   await crawler.init();
 
   // *** Get Result ***
-  await crawler.getList();
+  await crawler.getList({
+    pages: 3,
+    board: 'Beauty',
+  });
 
   // ***   Close   ***
   // await crawler.close();
-}
-
-runCrawler();
+})();
